refactor: remove chi as a dep for pure go impls
